refactor: replace arrow function component to functional component
