+++ src/navigation/index.tsx
@@ -86,7 +86,7 @@ export type RootStackParamList = {
 
 const Stack = createNativeStackNavigator<RootStackParamList>();
 
-const Navigation = () => {
+export default function Navigation() {
   const {isLoading, shouldRequestOnboarding} = useCheckOnboardingCompleted();
   const insets = useSafeAreaInsets();
 
@@ -227,7 +227,7 @@ const Navigation = () => {
       </NavigationContainer>
     </View>
   );
-};
+}
 
 const styles = StyleSheet.create({
   container: {
@@ -235,5 +235,3 @@ const styles = StyleSheet.create({
     backgroundColor: colors.OTBBlack,
   },
 });
-
-export default Navigation;
